api/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cytoolz==0.12.0
Django==4.1.3
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
drf-jwt==1.19.2
drf-yasg==1.21.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.4
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lru-dict==1.1.8
lxml==4.9.1
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.3
multitasking==0.0.11
netaddr==0.8.0
numpy==1.23.5
openapi-codec==1.3.2
packaging==21.3
pandas==1.5.2
parsimonious==0.8.1
patsy==0.5.3
protobuf==3.19.5
psycopg2==2.9.5
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
pymemcache==4.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-futures==1.0.0
rlp==2.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.9.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
statsmodels==0.13.5
stripe==5.4.0
threadpoolctl==3.1.0
toolz==0.12.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
varint==1.0.2
web3==5.31.3
websockets==9.1
yahooquery==2.2.15
yarl==1.8.2
